+++ src/components/Hero.tsx
@@ -7,13 +7,15 @@ interface HeroProps {
 
 export const Hero: React.FC<HeroProps> = ({ title, image }) => {
     return (
-        <div className="flex row justify-center items-center gap-4 h-screen">
+        <div className="flex flex-col justify-center items-center gap-4 ">
             <div className="w-4/12">
-                <h1 className="text-3xl font-bold">{title}</h1>
+                <h1 className="text-5xl font-bold w-[414px]   ">{title}</h1>
             </div>
-            <div className="w-4/12">
+            <div className="w-[812px] pl-[106px]">
                 <img src={image} width="100%" alt="Hero" loading="eager" />
             </div>
         </div>
     );
 };
+
+export type { HeroProps };