+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 03-07-2024
+### Added
+- Created a Page Generator Component
+    - created Results section to fillter tasks
+    - Created dynamic generator that can be easily modified
+    - added ReadME
+
 ## [1.0.3] - 03-07-2024
 ### Added
 - Created a Form Generator Component